- Changed Integer backing type to long long - Added 'current-time' primtive - Changed '+' and 'minus' to primitives (for speed) - Added timing utility functions to stdlib

+++ expression.h
@@ -10,7 +10,7 @@
 
 struct Null{};
 using Symbol = std::string;
-using Integer = int;
+using Integer = long long;
 using Boolean = bool;
 struct Pair;
 struct Closure;
